Updating crypto microbenchmarks
This commit adds crypto benchmarks that have been used in
bumblebench but were not commited before it was open sourced.
Also some crypto benchmarks are updated with more recent code.

+++ net/adoptopenjdk/bumblebench/crypto/EllipticCurveBench.java
@@ -50,7 +50,7 @@ public final class EllipticCurveBench extends MicroBench {
 
     protected long doBatch(long numIterations) throws InterruptedException {
         try {
-            for (int i = 0; i < numIterations; i++) {
+            for (long i = 0; i < numIterations; i++) {
                 keyPairGeneratora.initialize(new ECGenParameterSpec(curve), null);
                 keyPairGeneratorb.initialize(new ECGenParameterSpec(curve), null);
